Refs #35006 - Remove Pulp service classes
