+++ db/migrate/20200429153103_installed_package_bad_nvrea.rb
@@ -1,7 +1,7 @@
 class InstalledPackageBadNvrea < ActiveRecord::Migration[5.2]
   def up
     Katello::InstalledPackage.where(:epoch => "0").find_each do |pkg|
-      simple = Katello::Pulp::SimplePackage.new(pkg.attributes)
+      simple = Katello::SimplePackage.new(pkg.attributes)
       if pkg.nvrea != simple.nvrea
         pkg.update_column(:nvrea, simple.nvrea)
       end
