+++ test/models/concerns/host_managed_extensions_test.rb
@@ -231,7 +231,7 @@ module Katello
       super
       package_json = {:name => "foo", :version => "1", :release => "1.el7", :arch => "x86_64", :epoch => "1",
                       :nvra => "foo-1-1.el7.x86_64", :vendor => "Fedora"}
-      @foreman_host.import_package_profile([::Katello::Pulp::SimplePackage.new(package_json)])
+      @foreman_host.import_package_profile([::Katello::SimplePackage.new(package_json)])
       @nvra = 'foo-1-1.el7.x86_64'
       @foreman_host.reload
     end
@@ -244,7 +244,7 @@ module Katello
     end
 
     def test_import_package_profile_adds_removes_bulk
-      packages = [::Katello::Pulp::SimplePackage.new(:name => "betterfoo", :version => "1", :release => "1.el7",
+      packages = [::Katello::SimplePackage.new(:name => "betterfoo", :version => "1", :release => "1.el7",
                                                      :arch => "x86_64", :epoch => "1", :nvra => "betterfoo-1-1.el7.x86_64")]
       @foreman_host.import_package_profile(packages)
       assert_equal 1, @foreman_host.installed_packages.count
@@ -253,7 +253,7 @@ module Katello
       assert_equal '1', @foreman_host.installed_packages.first.epoch
 
       @foreman_host.reload
-      packages << ::Katello::Pulp::SimplePackage.new(:name => "alphabeta", :version => "1", :release => "2", :arch => "x86_64",
+      packages << ::Katello::SimplePackage.new(:name => "alphabeta", :version => "1", :release => "2", :arch => "x86_64",
                                                      :epoch => "1", :nvra => "alphabeta-1-2.x86_64")
       @foreman_host.import_package_profile(packages)
       assert_equal 2, @foreman_host.installed_packages.count
